+++ packages/components/pages/f-checkout/src/checkout-issues.js
@@ -73,6 +73,11 @@ const checkoutIssues = {
     DuplicateOrder: {
         shouldShowInDialog: true,
         shouldRedirectToMenu: false
+    },
+
+    BasketChanged: {
+        shouldShowInDialog: true,
+        shouldRedirectToMenu: true
     }
 };