+++ packages/components/pages/f-checkout/stories/Checkout.stories.js
@@ -34,6 +34,8 @@ const checkoutServerError = 'Checkout Error (Response from server is an error)';
 const placeOrderError = 'Place Order Duplicate Error (Response from server is an error)';
 const accessForbiddenError = 'Access Forbidden Get Checkout Error (Response from server is an error)';
 const getCheckoutError = 'Any other Get Checkout Error (Response from server is an error)';
+const invalidProductsError = 'Basket contains invalid products';
+const offlineProductsError = 'Basket contains offline products';
 const SERVER = 'SERVER';
 const accessForbiddenErrorCode = '403';
 const getCheckoutErrorCode = '500';
@@ -48,6 +50,8 @@ const timeNotAvailable = 'Selected time no longer available';
 const timeNotAvailableIssue = 'time-unavailable';
 const geolocationRequired = 'Geolocation required';
 const geolocationRequiredIssue = 'geolocation-required';
+const invalidProductsErrorCode = 'invalid-products';
+const offlineProductsErrorCode = 'offline-products';
 const serverTimeout = 'Server timeout';
 const serverTimeoutIssue = 'timeout';
 const duplicateIssue = 'duplicate';
@@ -83,6 +87,12 @@ const getCheckoutErrorOptions = {
     [getCheckoutError]: getCheckoutErrorCode,
     [noTimeAvailableError]: noTimeAvailable,
     [serverTimeout]: serverTimeoutIssue
+};
+
+const getBasketErrorOptions = {
+    None: null,
+    [invalidProductsError]: invalidProductsErrorCode,
+    [offlineProductsError]: offlineProductsErrorCode
 };
 
 const placeOrderErrorOptions = {
@@ -137,6 +147,10 @@ export const CheckoutComponent = () => ({
             default: select('Get Checkout Errors', getCheckoutErrorOptions, null)
         },
 
+        getBasketError: {
+            default: select('Get Basket Errors', getBasketErrorOptions, null)
+        },
+
         placeOrderError: {
             default: select('Place Order Errors', placeOrderErrorOptions)
         },
@@ -177,6 +191,11 @@ export const CheckoutComponent = () => ({
                     return `/checkout-${this.getCheckoutError}-get-error.json`;
                 }
             }
+
+            if (this.getBasketError) {
+                return `/get-basket-${this.getBasketError}.json`;
+            }
+
             return this.restriction ? `/get-basket-delivery-${this.restriction}.json` : `/get-basket-${this.serviceType}.json`;
         },