+++ packages/components/pages/f-checkout/CHANGELOG.md
@@ -3,12 +3,29 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-v3.10.0
+v3.12.0
 ------------------------------
 *November 26, 2021*
 
 ### Added
 - Ability to support Split Notes
+
+
+v3.11.0
+------------------------------
+*November 25, 2021*
+
+### Changed
+- `Error component` uses `BagSadBgIcon` from `f-vue-icons@3.0.0` and `f-card-with-content`.
+
+
+v3.10.0
+------------------------------
+*November 25, 2021*
+
+### Added
+- Error Dialog if Basket has changed.
+
 
 v3.9.0
 ------------------------------